Support more events at the same time
Doing persisting to ES and checking the DB for notifications and
processing them took a lot of time, and the previous setup meant that a
lot of events got dropped.

By using anotehr channel to push tasks onto, and starting 4 goroutines
that will take work off there to process, it's been able to handle
everything I throw at it. The buffered channel is key here.